requirements.txt
Django==2.2.16
djangorestframework==3.12.4
flake8==5.0.4
importlib-metadata==4.2.0
isort==5.10.1
mccabe==0.7.0
pycodestyle==2.9.1
pyflakes==2.5.0
pytz==2022.2.1
sqlparse==0.4.2
typing-extensions==4.3.0
zipp==3.8.1
